Some docstrings and some folders were renamed for a better understanding: scripts -> modules; modules -> scripts.
